add charge interaction code
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/AndreasKirsch@575 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/cuda/moleculeInteraction.h
@@ -70,7 +70,7 @@ public:
 		_componentDescriptorStorage.upload();
 	}
 
-	void calculate(float &potential, float &virial) {
+	void calculate(double &potential, double &virial) {
 		// pre-force calculation handling from CUDAForceCalculationComponents
 		_globalStats.preForceCalculation();
 		_moleculeStorage.preForceCalculation();
